Fix pod deletion for retry
