+++ workflow/controller/workflowpod_test.go
@@ -1570,36 +1570,6 @@ func TestPodMetadataWithWorkflowDefaults(t *testing.T) {
 	cancel()
 }
 
-func TestPodName(t *testing.T) {
-	nodeName := "nodename"
-
-	// short case
-	shortWfName := "wfname"
-	shortTemplateName := "templatename"
-
-	expected := fmt.Sprintf("%s-%s", shortWfName, shortTemplateName)
-	actual := ensurePodNamePrefixLength(expected)
-	assert.Equal(t, expected, actual)
-
-	name := PodName(shortWfName, nodeName, shortTemplateName)
-	assert.Equal(t, "wfname-templatename-1454367246", name)
-
-	// long case
-	longWfName := "alongworkflownamethatincludeslotsofdetailsandisessentiallyalargerunonsentencewithpoorstyleandnopunctuationtobehadwhatsoever"
-	longTemplateName := "alongtemplatenamethatincludessliightlymoredetailsandiscertainlyalargerunonstnencewithevenworsestylisticconcernsandpreposterouslyeliminatespunctuation"
-
-	sum := len(longWfName) + len(longTemplateName)
-	assert.Greater(t, sum, maxK8sResourceNameLength-k8sNamingHashLength)
-
-	expected = fmt.Sprintf("%s-%s", longWfName, longTemplateName)
-	actual = ensurePodNamePrefixLength(expected)
-
-	assert.Equal(t, maxK8sResourceNameLength-k8sNamingHashLength-1, len(actual))
-
-	name = PodName(longWfName, nodeName, longTemplateName)
-	assert.Equal(t, maxK8sResourceNameLength, len(name))
-}
-
 func TestPodExists(t *testing.T) {
 	cancel, controller := newController()
 	defer cancel()
